unknownsatoshi/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-environ==0.8.1
django-heroku==0.3.1
django-js-asset==1.2.2
djangoflutterwave==0.3.0
djangorestframework==3.12.4
Pillow==8.4.0
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
